requirements/base.txt
django==5.2.12
djangorestframework==3.15
django-jazzmin==3.0.4
psycopg2-binary
dj-database-url
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
Pillow
qrcode[pil]
resend
firebase-admin
python-decouple
django-filter
gunicorn
whitenoise
